:sparkles: Enhance ColorPicker with keyboard navigation and focus highlighting
This change introduces the `shouldEnableKeyboardHighlighting` prop to the `ColorPicker` component and its sub-components, enabling keyboard-only focus rings for interactive elements.

Users can now navigate and interact with the color area, sliders, preset colors, and more options inputs solely via the keyboard. The color area and preset colors support arrow key navigation, while other controls can be activated using `Enter` or `Spacebar`. Custom focus rings are applied using `useFocusRingPortal` and `useKeyboardFocusHighlighting` for improved accessibility.

+++ packages/color-picker/src/components/color-picker-popup/preset-colors/preset-button/PresetButton.tsx
@@ -1,5 +1,5 @@
-import { Icon } from '@chayns-components/core';
-import React, { useContext } from 'react';
+import { Icon, useFocusRingPortal, useKeyboardFocusHighlighting } from '@chayns-components/core';
+import React, { FocusEventHandler, useContext, useRef } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import type { IPresetColor } from '../../../../types/colorPicker';
 import { ColorPickerContext } from '../../../ColorPickerProvider';
@@ -10,16 +10,40 @@ interface PresetButtonProps {
     isCustom?: IPresetColor['isCustom'];
     onAdd?: (color: IPresetColor) => void;
     onRemove?: (id: IPresetColor['id']) => void;
+    shouldEnableKeyboardHighlighting?: boolean;
+    isKeyboardFocusable?: boolean;
+    onFocus?: FocusEventHandler<HTMLButtonElement>;
+    presetIndex?: number;
 }
 
-const PresetButton = ({ id, isCustom, onRemove, onAdd }: PresetButtonProps) => {
+const PresetButton = ({
+    id,
+    isCustom,
+    onRemove,
+    onAdd,
+    shouldEnableKeyboardHighlighting,
+    isKeyboardFocusable = false,
+    onFocus,
+    presetIndex,
+}: PresetButtonProps) => {
     const { selectedColor } = useContext(ColorPickerContext);
+    const buttonRef = useRef<HTMLButtonElement>(null);
+    const canRemove = Boolean(id && isCustom);
+    const shouldShowKeyboardHighlighting = useKeyboardFocusHighlighting(
+        shouldEnableKeyboardHighlighting,
+    );
 
-    const icon = isCustom ? ['fa-solid fa-trash'] : ['fa-solid fa-plus'];
+    useFocusRingPortal(buttonRef, {
+        isEnabled: shouldShowKeyboardHighlighting,
+        shape: 'circle',
+        padding: 3,
+    });
+
+    const icon = canRemove ? ['fa-solid fa-trash'] : ['fa-solid fa-plus'];
 
     const handleClick = () => {
-        if (id) {
-            if (typeof onRemove === 'function') {
+        if (canRemove) {
+            if (id && typeof onRemove === 'function') {
                 onRemove(id);
             }
         } else if (typeof onAdd === 'function' && selectedColor) {
@@ -28,8 +52,16 @@ const PresetButton = ({ id, isCustom, onRemove, onAdd }: PresetButtonProps) => {
     };
 
     return (
-        <StyledPresetButton onClick={handleClick} $isDisabled={!!id && !isCustom}>
-            <Icon icons={icon} size={isCustom ? 10 : 15} />
+        <StyledPresetButton
+            onClick={handleClick}
+            onFocus={onFocus}
+            ref={buttonRef}
+            data-preset-color-index={presetIndex}
+            tabIndex={isKeyboardFocusable ? 0 : -1}
+            type="button"
+            $isDisabled={false}
+        >
+            <Icon icons={icon} size={canRemove ? 10 : 15} />
         </StyledPresetButton>
     );
 };
